Fixed last_login and created_on declarations (via mah0001)

+++ ion_auth.sql
@@ -62,8 +62,8 @@ CREATE TABLE `users` (
   `activation_code` varchar(40) DEFAULT NULL,
   `forgotten_password_code` varchar(40) DEFAULT NULL,
   `remember_code` varchar(40) DEFAULT NULL,
-  `created_on` mediumint(11) unsigned NOT NULL,
-  `last_login` mediumint(11) unsigned NOT NULL,
+  `created_on` int(11) unsigned NOT NULL,
+  `last_login` int(11) unsigned NOT NULL,
   `active` int(1) unsigned DEFAULT NULL,
   PRIMARY KEY (`id`)
 );
